Rename nav_tree app to url_tree.
