+++ conman/pages/tests/test_handlers.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from conman.nav_tree.handlers import SimpleHandler
+from conman.url_tree.handlers import SimpleHandler
 from .. import handlers, views
