+++ conman/redirects/models.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from conman.nav_tree.models import Node
+from conman.url_tree.models import Node
 from . import handlers
 
 
@@ -11,5 +11,5 @@ class NodeRedirect(Node):
     This model holds the data required to make that connection.
     """
     handler = handlers.NodeRedirectHandler.path()
-    target = models.ForeignKey('nav_tree.Node', related_name='+')
+    target = models.ForeignKey('url_tree.Node', related_name='+')
     permanent = models.BooleanField(default=False, blank=True)
