+++ conman/redirects/tests/test_models.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from conman.nav_tree.tests.test_models import NODE_BASE_FIELDS
+from conman.url_tree.tests.test_models import NODE_BASE_FIELDS
 from .factories import ChildNodeRedirectFactory
 from ..models import NodeRedirect
