+++ conman/redirects/tests/test_views.py
@@ -1,7 +1,7 @@
 from django.test import TestCase
 
-from conman.nav_tree.tests.factories import ChildNodeFactory
 from conman.tests.utils import RequestTestCase
+from conman.url_tree.tests.factories import ChildNodeFactory
 from .factories import ChildNodeRedirectFactory
 from .. import views
