+++ example/example/urls.py
@@ -3,5 +3,5 @@ from django.conf.urls import include, url
 
 urlpatterns = [
     url(r'^cms/', include('conman.cms.urls')),
-    url(r'', include('conman.nav_tree.urls')),
+    url(r'', include('conman.url_tree.urls')),
 ]
